requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
bcrypt==4.1.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httptools==0.6.1
idna==3.6
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
sniffio==1.3.0
starlette==0.35.1
typing_extensions==4.9.0
uvicorn==0.26.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
